requirements.txt
# Core AI Agent Dependencies
google-generativeai>=0.8.0

# Data Processing & Storage
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Observability
structlog>=24.0.0

# Async Support
aiofiles>=23.0.0
httpx>=0.25.0

# Vector Database for Knowledge Retrieval
chromadb>=0.4.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0

# Mock integrations
requests>=2.31.0